requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
